# backend/src/main/kotlin/tech/bhrigu/almira/reminder/ReminderController.kt
package tech.bhrigu.almira.reminder

import jakarta.validation.Valid
import jakarta.validation.constraints.NotBlank
import org.springframework.http.HttpStatus
import org.springframework.http.ResponseEntity
import org.springframework.web.bind.annotation.DeleteMapping
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.ResponseStatus
import org.springframework.web.bind.annotation.RestController
import tech.bhrigu.almira.common.IndianNumbers
import tech.bhrigu.almira.investment.InvestmentService
import java.math.BigDecimal
import java.time.LocalDate
import java.time.temporal.ChronoUnit
import java.util.UUID

data class CreateReminderBody(
    val investmentId: UUID? = null,
    val liabilityId: UUID? = null,
    val kind: String = "custom",
    @field:NotBlank(message = "What should we remind you about?")
    val title: String,
    val dueDate: LocalDate,
    val leadDays: Int = 7,
    val recurrence: String = "none",
    val amount: BigDecimal? = null,
    val note: String? = null,
)

data class SnoozeBody(val until: LocalDate)

data class ReminderResponse(
    val id: UUID,
    val kind: String,
    val title: String,
    val recordTitle: String?,
    val investmentId: UUID?,
    val liabilityId: UUID?,
    val dueDate: LocalDate,
    /** Snoozed date if there is one, otherwise the due date — what to sort by. */
    val effectiveDate: LocalDate,
    val daysAway: Long,
    val leadDays: Int,
    val recurrence: String,
    val amount: BigDecimal?,
    val amountFormatted: String?,
    val status: String,
    val note: String?,
    val autoGenerated: Boolean,
)

data class CompleteReminderResponse(
    val reminder: ReminderResponse?,
    /** Present for a maturity: the moment to offer the renewal. */
    val rollover: RolloverDraft?,
)

@RestController
@RequestMapping("/api/v1/households/{householdId}/reminders")
class ReminderController(
    private val service: ReminderService,
    private val investments: InvestmentService,
) {

    @GetMapping
    fun list(
        @PathVariable householdId: UUID,
        @RequestParam(required = false) withinDays: Int?,
        @RequestParam(defaultValue = "false") includeDone: Boolean,
    ): List<ReminderResponse> =
        service.list(householdId, withinDays, includeDone).map { it.toResponse() }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    fun create(
        @PathVariable householdId: UUID,
        @RequestBody @Valid body: CreateReminderBody,
    ): ReminderResponse = service.create(
        householdId, body.investmentId, body.liabilityId, body.kind, body.title,
        body.dueDate, body.leadDays, body.recurrence, body.amount, body.note,
    ).toResponse()

    @PostMapping("/{id}/snooze")
    fun snooze(
        @PathVariable householdId: UUID,
        @PathVariable id: UUID,
        @RequestBody @Valid body: SnoozeBody,
    ): ReminderResponse = service.snooze(householdId, id, body.until).toResponse()

    @PostMapping("/{id}/done")
    fun complete(
        @PathVariable householdId: UUID,
        @PathVariable id: UUID,
    ): CompleteReminderResponse {
        // The linked holding is fetched through the caller's own visibility, so
        // a rollover is only ever offered for something they can actually see.
        val reminder = service.list(householdId, null, true).firstOrNull { it.id == id }
        val investment = reminder?.investmentId?.let {
            runCatching { investments.get(householdId, it) }.getOrNull()
        }
        val result = service.complete(householdId, id, investment)
        return CompleteReminderResponse(result.reminder?.toResponse(), result.rollover)
    }

    @DeleteMapping("/{id}")
    fun delete(@PathVariable householdId: UUID, @PathVariable id: UUID): ResponseEntity<Void> {
        service.delete(householdId, id)
        return ResponseEntity.noContent().build()
    }

    private fun ReminderRow.toResponse() = ReminderResponse(
        id = id, kind = kind, title = title, recordTitle = recordTitle,
        investmentId = investmentId, liabilityId = liabilityId,
        dueDate = dueDate, effectiveDate = effectiveDate,
        daysAway = ChronoUnit.DAYS.between(LocalDate.now(), effectiveDate),
        leadDays = leadDays, recurrence = recurrence, amount = amount,
        amountFormatted = amount?.let { IndianNumbers.rupees(it) },
        status = status, note = note, autoGenerated = autoGenerated,
    )
}