# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "Cashflow": {
        "properties": {
          "from": {
            "format": "date",
            "type": "string"
          },
          "months": {
            "items": {
              "$ref": "#/components/schemas/CashflowMonth"
            },
            "type": "array"
          },
          "until": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "from",
          "months",
          "until"
        ],
        "type": "object"
      },
      "CashflowEntry": {
        "properties": {
          "amount": {
            "type": "number"
          },
          "amountFormatted": {
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "recordId": {
            "format": "uuid",
            "type": "string"
          },
          "recordType": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "amountFormatted",
          "date",
          "direction",
          "kind",
          "recordId",
          "recordType",
          "title"
        ],
        "type": "object"
      },
      "CashflowMonth": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/CashflowEntry"
            },
            "type": "array"
          },
          "inflow": {
            "type": "number"
          },
          "inflowFormatted": {
            "type": "string"
          },
          "month": {
            "type": "string"
          },
          "net": {
            "type": "number"
          },
          "netFormatted": {
            "type": "string"
          },
          "out": {
            "type": "number"
          },
          "outFormatted": {
            "type": "string"
          }
        },
        "required": [
          "entries",
          "inflow",
          "inflowFormatted",
          "month",
          "net",
          "netFormatted",
          "out",
          "outFormatted"
        ],
        "type": "object"
      },
...
      "CompleteReminderResponse": {
        "properties": {
          "reminder": {
            "$ref": "#/components/schemas/ReminderResponse"
          },
          "rollover": {
            "$ref": "#/components/schemas/RolloverDraft"
          }
        },
        "type": "object"
      },
...
      "CreateReminderBody": {
        "properties": {
          "amount": {
            "type": "number"
          },
          "dueDate": {
            "format": "date",
            "type": "string"
          },
          "investmentId": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "leadDays": {
            "format": "int32",
            "type": "integer"
          },
          "liabilityId": {
            "format": "uuid",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "recurrence": {
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "dueDate",
          "kind",
          "leadDays",
          "recurrence",
          "title"
        ],
        "type": "object"
      },
...
      "HouseholdResponse": {
        "properties": {
          "baseCurrency": {
            "type": "string"
          },
          "defaultVisibility": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "memberCount": {
            "format": "int32",
            "type": "integer"
          },
          "myMemberId": {
            "format": "uuid",
            "type": "string"
          },
          "myRole": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "baseCurrency",
          "defaultVisibility",
          "id",
          "memberCount",
          "myRole",
          "name",
          "version"
        ],
        "type": "object"
      },
...
      "NetWorthTrend": {
        "properties": {
          "basis": {
            "type": "string"
          },
          "historyBeginsAt": {
            "format": "date",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array"
          },
          "scope": {
            "type": "string"
          }
        },
        "required": [
          "basis",
          "points",
          "scope"
        ],
        "type": "object"
      },
...
      "ReminderResponse": {
        "properties": {
          "amount": {
            "type": "number"
          },
          "amountFormatted": {
            "type": "string"
          },
          "autoGenerated": {
            "type": "boolean"
          },
          "daysAway": {
            "format": "int64",
            "type": "integer"
          },
          "dueDate": {
            "format": "date",
            "type": "string"
          },
          "effectiveDate": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "investmentId": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "leadDays": {
            "format": "int32",
            "type": "integer"
          },
          "liabilityId": {
            "format": "uuid",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "recordTitle": {
            "type": "string"
          },
          "recurrence": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "autoGenerated",
          "daysAway",
          "dueDate",
          "effectiveDate",
          "id",
          "kind",
          "leadDays",
          "recurrence",
          "status",
          "title"
        ],
        "type": "object"
      },
...
      "RolloverDraft": {
        "properties": {
          "investmentId": {
            "format": "uuid",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "suggestedAmount": {
            "type": "number"
          },
          "suggestedStartDate": {
            "format": "date",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "typeId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "investmentId",
          "message",
          "suggestedStartDate",
          "title",
          "typeId"
        ],
        "type": "object"
      },
      "SearchHit": {
        "properties": {
          "amount": {
            "type": "number"
          },
          "amountFormatted": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "route": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "route",
          "title"
        ],
        "type": "object"
      },
      "SearchResults": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array"
          },
          "investments": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array"
          },
          "liabilities": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array"
          },
          "query": {
            "type": "string"
          },
          "total": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "accounts",
          "documents",
          "investments",
          "liabilities",
          "people",
          "query",
          "total"
        ],
        "type": "object"
      },
...
      "SnoozeBody": {
        "properties": {
          "until": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "until"
        ],
        "type": "object"
      },
...
      "TrendPoint": {
        "properties": {
          "assets": {
            "type": "number"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "liabilities": {
            "type": "number"
          },
          "netWorth": {
            "type": "number"
          },
          "netWorthFormatted": {
            "type": "string"
          }
        },
        "required": [
          "assets",
          "date",
          "liabilities",
          "netWorth",
          "netWorthFormatted"
        ],
        "type": "object"
      },
...
      "UpdateHouseholdBody": {
        "properties": {
          "defaultVisibility": {
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Access token from POST /api/v1/auth/otp/verify.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "padmini.koduri@bhrigu.tech",
      "name": "Almira"
    },
    "description": "Every rupee a family has invested or owes, in one private place.\n\n**Authentication.** Phone OTP issues a short-lived access token and a\nrotating refresh token. Send `Authorization: Bearer <accessToken>`.\nRefresh tokens are single-use: presenting one twice is treated as theft\nand ends the session.\n\n**Privacy.** Every read passes a per-record visibility check enforced by\nPostgreSQL row-level security. A record you may not see returns 404, not\n403 — a 403 would confirm it exists. Totals are computed through the same\nfilter, so a private record contributes nothing, not even its amount, to\nanother member's figures.\n\n**Money** is decimal, never floating point. Amounts also arrive formatted\n(`₹1,76,875`) so every surface renders them identically.\n\n**Concurrency.** Writes carry the record's `version`; a stale write is\nrejected with 409 and the current version. Creates accept a client-supplied\n`id`, so an offline capture keeps its identity and a retry is idempotent.\n\n**Stability.** v1 is frozen and additive-only. Breaking changes go to v2.",
    "title": "Almira API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/households/{householdId}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/reminders": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "withinDays",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "includeDone",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReminderResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "reminder-controller"
        ]
      },
      "post": {
        "operationId": "create_1",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReminderBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "reminder-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/reminders/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "reminder-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/reminders/{id}/done": {
      "post": {
        "operationId": "complete",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteReminderResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "reminder-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/reminders/{id}/snooze": {
      "post": {
        "operationId": "snooze",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnoozeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "reminder-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/reports/cashflow": {
      "get": {
        "operationId": "cashflow",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 3,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cashflow"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "reports-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/reports/net-worth-trend": {
      "get": {
        "operationId": "netWorthTrend",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 12,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "household",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "member",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorthTrend"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "reports-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/search": {
      "get": {
        "operationId": "search",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResults"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "search-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ]
}